requirements.txt
annotated-types==0.5.0
anyascii==0.3.2
asgiref==3.7.2
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.11.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==41.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
Django==4.2.4
django-allauth==0.54.0
django-bootstrap-datepicker-plus==5.0.4
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-rest-swagger==2.2.0
django-silk==5.0.3
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
drf-yasg==1.21.7
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.2
filetype==1.2.0
firebase-admin==6.2.0
gcloud==0.18.3
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
gprof2dot==2022.7.29
grpcio==1.57.0
grpcio-status==1.57.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jwcrypto==1.5.0
l18n==2021.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msgpack==1.0.5
oauth2client==4.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.8.23
platformdirs==3.10.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-jwt==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
whitenoise==6.6.0
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
telepath==0.3.1
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.3
wagtail==5.0.2
wcwidth==0.2.6
webencodings==0.5.1
Willow==1.5.1
wrapt==1.15.0
gunicorn==21.2.0